Реализовано мввм первого экрана
