+++ Moovies/Moovies/Helpers/UpdateViewDataTypealias.swift
@@ -1,0 +1,10 @@
+//
+//  UpdateViewDataTypealias.swift
+//  Moovies
+//
+//  Created by Vera Zaitseva on 11.10.2021.
+//
+
+import Foundation
+
+typealias UpdateMovieHandler = (([MovieData]) -> Void)
