+++ Moovies/Moovies/Helpers/VoidHandler.swift
@@ -1,0 +1,10 @@
+//
+//  VoidHandler.swift
+//  Moovies
+//
+//  Created by Vera Zaitseva on 11.10.2021.
+//
+
+import Foundation
+
+typealias VoidHandler = (() -> Void)
