+++ Moovies/Moovies/MovieDescriptionModule/Model/MoovieDescription.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct Description: Decodable {
-     var posterPath: String?
-     let title: String
-     let overview: String
- }
+    var posterPath: String?
+    let title: String
+    let overview: String
+}
